requirements.txt
google-genai>=1.0.0
edge-tts>=6.1.0
pydub>=0.25.0
python-dotenv>=1.0.0
rich>=13.0.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.14
flask>=3.0.0
flask-cors>=4.0.0
pyinstaller>=6.0.0
